// src/lib/services/models/RequestSummaryChitStandardResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { HttpStatusCode } from './HttpStatusCode';
import type { RequestSummary } from './RequestSummary';
export type RequestSummaryChitStandardResponse = {
    href?: string | null;
    relations?: Array<string> | null;
    method?: string | null;
    routeName?: string | null;
    routeValues?: any;
    status?: boolean;
    message?: string | null;
    data?: RequestSummary;
    statusCode?: HttpStatusCode;
    errors?: any;
};