// src/lib/services/models/RequestViewPagedCollectionChitStandardResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { HttpStatusCode } from './HttpStatusCode';
import type { RequestViewPagedCollection } from './RequestViewPagedCollection';
export type RequestViewPagedCollectionChitStandardResponse = {
    href?: string | null;
    relations?: Array<string> | null;
    method?: string | null;
    routeName?: string | null;
    routeValues?: any;
    status?: boolean;
    message?: string | null;
    data?: RequestViewPagedCollection;
    statusCode?: HttpStatusCode;
    errors?: any;
};